More slides for appendix
Including these in the appendix in case someone (like Lester) asks about the Bollinger bands. As a reminder, we wanted to look at an alternative to key word correlation to stock price in case it didn't work. These slides show the images only of recession to stock price on the top and the stock's Bollinger band on the bottom. There doesn't appear to be any relationship between key word/stock price and stock volatility measured with Bollinger bands.